requirements.txt
Django==5.2.5
djangorestframework==3.16.1
pytest==8.4.2
pytest-django==4.11.1